refactor: add SDK_INFO to core; simplify SDK resource creation
