+++ packages/opentelemetry-base/package.json
@@ -3,6 +3,10 @@
   "version": "0.4.0",
   "description": "OpenTelemetry base provides base code for the SDK packages",
   "main": "build/src/index.js",
+  "browser": {
+    "./src/platform/index.ts": "./src/platform/browser/index.ts",
+    "./build/src/platform/index.js": "./build/src/platform/browser/index.js"
+  },
   "types": "build/src/index.d.ts",
   "repository": "open-telemetry/opentelemetry-js",
   "scripts": {
