+++ packages/opentelemetry-base/src/platform/node/constants.ts
@@ -1,5 +1,5 @@
-/*!
- * Copyright 2019, OpenTelemetry Authors
+/**
+ * Copyright 2020, OpenTelemetry Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,5 +14,12 @@
  * limitations under the License.
  */
 
-export * from './ExportResult';
-export * from './platform';
+import { VERSION } from '../../version';
+
+/** Constants describing the SDK in use */
+export const SDK_INFO = {
+  NAME: 'opentelemetry',
+  RUNTIME: 'node',
+  LANGUAGE: 'nodejs',
+  VERSION: VERSION,
+};
