+++ packages/opentelemetry-node/src/NodeTracerProvider.ts
@@ -19,7 +19,6 @@ import {
   BasicTracerProvider,
   SDKRegistrationConfig,
 } from '@opentelemetry/tracing';
-import { Resource, SDK_INFO } from '@opentelemetry/resources';
 import { DEFAULT_INSTRUMENTATION_PLUGINS, NodeTracerConfig } from './config';
 import { PluginLoader } from './instrumentation/PluginLoader';
 
@@ -31,9 +30,6 @@ import { PluginLoader } from './instrumentation/PluginLoader';
  * @param config Configuration object for SDK registration
  */
 export class NodeTracerProvider extends BasicTracerProvider {
-  readonly resource: Resource = Resource.createTelemetrySDKResource(
-    SDK_INFO.PLATFORM_NODE
-  );
   private readonly _pluginLoader: PluginLoader;
 
   /**
