+++ packages/opentelemetry-resources/package.json
@@ -54,6 +54,7 @@
     "typescript": "3.7.2"
   },
   "dependencies": {
-    "@opentelemetry/api": "^0.4.0"
+    "@opentelemetry/api": "^0.4.0",
+    "@opentelemetry/base": "^0.4.0"
   }
 }
