+++ packages/opentelemetry-resources/src/constants.ts
@@ -1,21 +1,3 @@
-/*!
- * Copyright 2020, OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { VERSION } from './version';
-
 /**
  * Copyright 2020, OpenTelemetry Authors
  *
@@ -140,12 +122,4 @@ export const SERVICE_RESOURCE = {
 
   /** The version string of the service API or implementation. */
   VERSION: 'service.version',
-};
-
-/** Constants describing the SDK in use  */
-export const SDK_INFO = {
-  NAME: 'opentelemetry',
-  VERSION: VERSION,
-  PLATFORM_NODE: 'nodejs',
-  PLATFORM_WEB: 'webjs',
 };
